Move ldapclient class to osuosl colo only
Currently we're enabling ldap for auth on all our puppet'd hosts,
regardless of whether they can actually talk to our LDAP
servers. That, obviously, is kinda silly. Long-term, we should work up
a way for non-OSUOSL hosts to be able to talk to LDAP, but for now,
moving the ldapclient class inclusion into the osuosl.yaml colo file
gets us the result we need.